packages/ekho_core/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ekho-core"
version = "0.1.0"
description = "Core utilities for EKHO voice translation platform"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.0",
    "pydub>=0.25.1",
    "ffmpeg-python>=0.2.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "black>=24.0",
    "ruff>=0.6",
    "isort>=5.13",
]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=ekho_core --cov-report=term-missing"
